% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detec3DRelation.R
\name{Detecte3DRelation}
\alias{Detecte3DRelation}
\title{A 3D Object relationship detection function}
\usage{
Detecte3DRelation(TableFrom, TableTo, ClosedFrom, ClosedTo)
}
\arguments{
\item{TableFrom, }{TableTo are tables with; ClosedFrom, ClosedTo}
}
\value{
A table of combined the input tables with result
}
\description{
This function allows you to compare objects saved in two list
}
\keyword{3D}
